Update patch set 4
Patch Set 4: -Verified

No problem. This line will be removed. Too bad it is not allowed because it is rewritable in the misc partition and anyone can do it with root access. Google and LG should have thought this through before they release this smartphone and the code at a worldwide level. Not to mention that WiFi Mac address is random at android 10 while we maintain the real Mac.
One question.
Misc partition CAN be wiped. And misc partition was involved a couple of years ago in a twrp mistake and got wiped in all devices.
Of course you all know here in lineage os that when we reflashed kdz to unbrick our device, our Mac address was changed ???
So even if we sent it to LG for a repair instead, it'd come back with a different WiFi Mac address.
What is not allowed then ??? To repair people device ???

As a developer I have to protect bricked devices. And if someone wipe misc partition , I have to offer him a fix. So this is the fix. Otherwise he has to flash kdz and yet his Mac address will change again.
Blame LG. But this thing is the real shortcut fix to avoid kdz and it can't be more official.
If you still -2 then tomorrow I will remove the line and let bricked devices flash kdz.

Patch-set: 4
Reviewer: Gerrit User 18258 <18258@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0