Update patch set 3
Abandoned

it was a great idea, but writting with a service to root partition is not possible. Also writting to persist partition is not possible because persist is not mounted at that time. Too much trouble for nothing. Maybe in the future i'll do this one.

Patch-set: 3
Status: abandoned
Tag: autogenerated:gerrit:abandon