Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> (8 comments)

Most of these fixed here:
https://review.lineageos.org/c/LineageOS/android_device_lge_g4-common/+/285401

Mike thank you for observations, but i can't correct the code of others inside their commit. When you cherry-pick from google sources for example, you edit the commit and alter it ? No.
So i cherry-pick the stuff as they are, maybe add couple of lines if i squash, i commit and then i do another commit of mine if something needs a correction. Those stuff now that were downvoted they are already merged in the lineage since months or years. 

Guys just merge the commits and don't put me on hold for years because of a gap or whatever. The important thing is the code to work. If it works, merge and add me an observation so i can come back later and improve it. You know there are commits right now on hold since 2018 ?
That's unacceptable. Do you want to have LG G4 officially on your server or not ? Unofficially i'm already on android 10 stable with only camera not working (until i fix it) and your official is still down to MarshMallow and N is not even stable enough. So if you want this device to come up official, do not let commits on hold for years. Otherwise i don't care. i'll abandon it AGAIN and keep doing it unofficially. Thank you in advanced.

Patch-set: 3
Reviewer: Gerrit User 18258 <18258@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>